fix: Makes `draft` optional for scorecards (#46)
We document the `draft` property of scorecards as optional but don't
provide a default for the resource. This updates the resource schema to
mark draft as both optional and computed (you need both to avoid
conflicts).

It also fixes the README which mistakenly pointed users to
`cortex/cortex` instead of the correct provider name of
`cortexapp/cortex`.